+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to the "SQLite IntelliView" extension will be documented in this file.
 
+## [Unreleased]
+
+### 🐛 Bug Fixes
+
+- Fixed issue where the data tab would open with 1000 results instead of the intended 100.
+
 ## [0.2.11] - 2025-07-07
 
 ### ⚠️ Beta Warning